rename some functions to be more clear about what they do. rewrite some docstrings to be more clear about what the function does. Move some functions around, so they're in better file names. eg, break utils into merging_dataframes, normalize_dataframes, and actual non-pandas utils
