+++ .codeclimate.yml
@@ -8,7 +8,7 @@ checks:
       threshold: 3
   file-lines:
     config:
-      threshold: 254
+      threshold: 200
   method-complexity:
     config:
       threshold: 9
