+++ merging_code/scrape_geocodes.py
@@ -15,9 +15,10 @@ import ssl
 import certifi
 import geopy
 
+from merging_code.merge_dataframes import get_dataframe_from_merged_table_metadata
+from merging_code.normalize_dataframes import add_empty_columns
 from merging_code.secrets import GEOCODE_API_KEY
 from merging_code.utils import get_dict_from_json_file, write_dict_to_json_file
-from merging_code.utils import get_dataframe_from_merged_table_metadata, add_empty_columns
 from file_locations import CENSUS_FINAL_CSV_FILENAME, FBI_CRIME_COMBINED_CSV_FILENAME, EXPERIAN_FINAL_CSV_FILENAME
 from file_locations import GEOCODE_CACHED_JSON_FILENAME, GEOCODE_FINAL_CSV_FILENAME
 
@@ -35,13 +36,6 @@ CSV_FILES_TO_MERGE = [{
 }]
 
 
-def get_census_cities_and_states_dataframe():
-  """ Load the pandas dataframe from table metadata. """
-  dataframe = get_dataframe_from_merged_table_metadata(CSV_FILES_TO_MERGE)
-  dataframe = dataframe[['city', 'state']]
-  return dataframe
-
-
 def get_geopy_googlev3_locator(geocode_api_key):
   """ Get the geopy geolocator object, setup with goog auth. """
   ctx = ssl.create_default_context(cafile=certifi.where())
@@ -56,8 +50,7 @@ def get_geopy_googlev3_locator(geocode_api_key):
   return geolocator
 
 
-def set_geo_metadata_to_dict(cached_json, location, reverse_address,
-                             search_query):
+def cache_geo_metadata(cached_json, location, reverse_address, search_query):
   """ Set single city geo metadata values into the main cache dict, and return the dict. """
   cached_json[search_query] = {
     'latitude': location.latitude,
@@ -75,14 +68,15 @@ def get_reverse_address(geolocator, location):
   return reverse.address
 
 
-def set_geo_metadata_to_dataframe_row(row, location, reverse_address):
+def add_geo_metadata_to_row(row, location, reverse_address):
   """ Overwrite empty cell values in a dataframe, to add geo metadata """
   row['latitude'] = location.latitude
   row['longitude'] = location.longitude
   row['reverse_address'] = reverse_address
+  return row
 
 
-def set_geo_metadata_to_dataframe(dataframe):
+def add_geo_metadata_to_dataframe(dataframe):
   """ Iterate over all cities in the dataframe, then add geo metadata to all of them """
   geolocator = get_geopy_googlev3_locator(GEOCODE_API_KEY)
   cached_json = get_dict_from_json_file(GEOCODE_CACHED_JSON_FILENAME)
@@ -99,9 +93,8 @@ def set_geo_metadata_to_dataframe(dataframe):
       continue
     location = geolocator.geocode(search_query)
     reverse_address = get_reverse_address(geolocator, location)
-    set_geo_metadata_to_dataframe_row(row, location, reverse_address)
-    set_geo_metadata_to_dict(cached_json, location, reverse_address,
-                             search_query)
+    row = add_geo_metadata_to_row(row, location, reverse_address)
+    cache_geo_metadata(cached_json, location, reverse_address, search_query)
     api_count += 2  # two api hits per loop, 1 for geocode, 1 for reverse address
     if api_count % 50 == 0:
       log_msg = '### api_count: {}, cache_count: {} ###'.format(
@@ -109,12 +102,14 @@ def set_geo_metadata_to_dataframe(dataframe):
       print(log_msg)
       write_dict_to_json_file(GEOCODE_CACHED_JSON_FILENAME, cached_json)
   write_dict_to_json_file(GEOCODE_CACHED_JSON_FILENAME, cached_json)
+  return dataframe
 
 
 def get_final_dataframe():
   """ The main function which returns the final dataframe. """
-  dataframe = get_census_cities_and_states_dataframe()
-  set_geo_metadata_to_dataframe(dataframe)
+  dataframe = get_dataframe_from_merged_table_metadata(CSV_FILES_TO_MERGE)
+  dataframe = dataframe[['city', 'state']]
+  dataframe = add_geo_metadata_to_dataframe(dataframe)
   dataframe = dataframe.sort_values(by=['state', 'city'])
   return dataframe
