+++ requirements-dev.txt
@@ -1,5 +1,6 @@
 -r requirements.txt
 certifi==2019.3.9
+coloredlogs==10.0
 flake8==3.7.7
 fuzzywuzzy==0.17.0
 geopy==1.20.0
